+++ apps/modern-component-data-fetch/provider/src/components/ClientDowngrade/index.data.ts
@@ -1,0 +1,15 @@
+export type Data = {
+  data: string;
+};
+
+export const fetchData = async (): Promise<Data> => {
+  throw new Error('force downgrade!');
+
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve({
+        data: `fetch data from provider \n${new Date()}`,
+      });
+    }, 1000);
+  });
+};